refactor: 파일명칭 변경 및 scss 수정
