+++ frontend/src/pages/AllContainer.tsx
@@ -8,7 +8,7 @@ import {
   getContainers,
   getContainerMembers,
   leaveContainer,
-} from "../api/home";
+} from "../api/homeApi";
 import type {
   ContainerResponseDto,
   GroupMemberResponseDto,
