+++ frontend/src/pages/SharedContainer.tsx
@@ -8,7 +8,7 @@ import {
   getSharedContainers,
   getContainerMembers,
   leaveContainer,
-} from "../api/home";
+} from "../api/homeApi";
 import type {
   ContainerResponseDto,
   GroupMemberResponseDto,
