+++ frontend/src/styles/AppSidebar.module.scss
@@ -147,7 +147,7 @@
   color: #fff;
 
   &:hover {
-    background-color: #0071db;
+    background-color: #0085ff;
   }
 
   :global(body.dark-mode) & {
